modul_1_Python_Application_Development/09_Backend-разработка_FastAPI_для_backend-сервера/09_lection/requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
psycopg2-binary==2.9.5
loguru==0.6.0
requests==2.28.2